% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cart.r2}
\alias{cart.r2}
\title{Create a cartogram SpatialPolygonsDataFrame}
\usage{
cart.r2(cart, var)
}
\arguments{
\item{cart}{The cartogram \code{SpatialPolygonsDataFrame}}

\item{var}{The cartogram variable}
}
\value{
The R-squared value
}
\description{
Given a \code{SpatialPolygonsDataFrame} as a cartogram
and the cartogram variable,  computes the R-squared obtained from regressing
cartogram zone areas on the cartogram variable,  without an intercept
}
\examples{
data(georgia)
georgia.carto <- quick.carto(georgia2,georgia2$TotPop90)
cart.r2(georgia.carto,georgia2$TotPop90)

# Investigate performance as resolution changes
perf <- function(r) cart.r2(quick.carto(georgia2,res=64*r,georgia2$TotPop90),
     georgia2$TotPop90)
score <- rep(0,8)
for (i in 1:8) score[i] <- perf(i)
plot(64*(1:8),score,xlab='Resolution',ylab='Cartogram Correlation',type='b')

# Suppose we start with an inappropriate map projection?
poor.carto <- quick.carto(georgia,georgia$TotPop90)
georgia.carto <- quick.carto(georgia2,georgia2$TotPop90)
cart.r2(poor.carto,georgia$TotPop90)
cart.r2(georgia.carto,georgia2$TotPop90)
}
\keyword{cartogram}

