% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cart.resids.R
\name{cart.resids}
\alias{cart.resids}
\title{Create a cartogram SpatialPolygonsDataFrame}
\usage{
cart.resids(cart, var)
}
\arguments{
\item{cart}{The cartogram \code{SpatialPolygonsDataFrame}}

\item{var}{The cartogram variable}
}
\value{
The residuals (in cartogram area units) for each zone
}
\description{
Given a \code{SpatialPolygonsDataFrame} as a cartogram
and the cartogram variable,  computes the residual between the
desired area and the actual area for each zone in cartogram area units
}
\examples{
data(georgia)
georgia.carto <- quick.carto(georgia2,georgia2$TotPop90)
discrep <- cart.resids(georgia.carto,georgia2$TotPop90)
shs <- auto.shading(c(discrep,-discrep),n=5,col=brewer.pal(5,'RdBu'))
choropleth(georgia.carto,discrep,shading=shs)
choro.legend(105,105,shs)
}
\keyword{cartogram}

