% Generated by roxygen2 (4.0.2): do not edit by hand
\name{carto.transform}
\alias{carto.transform}
\title{Create a warping function to apply to Spatial* and Spatial*DataFrame objects}
\usage{
carto.transform(spdf, v, extend = 0.05, res = 128, index = 1,
  thresh = 0.1, blur = 0, prec = 0.5, gapdens = 1)
}
\arguments{
\item{spdf}{A \code{SpatialPolygonsDataFrame}}

\item{v}{The variable providing the values - typically population counts}

\item{extend}{The extent beyond the limits of the \code{SpatialPolygonsDataFrame} to estimate the warping as a proportion of span}

\item{res}{The resolution of the grid for the warping - default 128}

\item{index}{Index of data item to be used for cartogram - only used if \code{v} is missing}

\item{thresh}{Lowest density value allowed - will replace pixels with \code{max(value,thresh*mean(value))} - 0.1 works well}

\item{blur}{Degree of Gaussian blur to apply to density grid,  prior to computing the cartogram transform}

\item{prec}{The precision when carrying out interpolation, in units of the warping grid - 0.5 works well}

\item{gapdens}{The density to assign to gaps between the polygons - relative to average density.}
}
\value{
a function taking a \code{Spatial*} or \code{Spatial*DataFrame} object,  returning a warped version of the object
}
\description{
Create a cartogram transform given a \code{SpatialPolygons*} object and a variable
The returned value is a function that performs the transform.
}
\examples{
# Make a cartogram for Newhaven census blocks
require(GISTools)
data(newhaven)
to.carto <- carto.transform(blocks,blocks$POP1990,thresh=0.1)
# Create the transform function
# Plot the blocks carto
plot(to.carto(blocks))
# Add roads,  transformed to cartogram space
plot(to.carto(roads),add=TRUE,col='lightgrey')
# Add forced entry residential burglaries, transformed to cartogram space
plot(to.carto(burgres.f),add=TRUE,col='red',pch=16)
}
\references{
A Diffusion-based method for producing density equalizing maps, Michael T. Gastner and M. E. J. Newman, Proc. Natl. Acad. Sci. USA 101, 7499-7504 (2004)
}
\seealso{
Other transformers: \code{\link{quick.carto.transform}};
  \code{\link{warp.lines}}; \code{\link{warp.points}};
  \code{\link{warp.polys}}
}
\keyword{cartogram}

