% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dispersion.R
\name{dispersion}
\alias{dispersion}
\title{Visual representation of cartogram transformation}
\usage{
dispersion(spdf, add = FALSE, col = "red")
}
\arguments{
\item{spdf}{A \code{Spatial*} object created via the \code{quick.carto} function, either directly or indirectly via \code{carto.transform}}

\item{add}{Whether to add the graphic to an existing plot}

\item{col}{Colour of the graphic}
}
\description{
Given a transformed Spatial object draws the dispersion cloud
}
\examples{
# Load an example SpatialPolygonsDataFrame from GISTools
data(georgia)
# Create the cartogram.  TotPop90 contains 1990 county populations
# "georgia2" is used as it has a plane projection coordinate system
georgia.carto <- quick.carto(georgia2,georgia2$TotPop90)
# Draw the dispersion cloud
dispersion(georgia.carto)
# Superimpose the transformed map
plot(georgia.carto,add=TRUE)
}
\references{
A Diffusion-based method for producing density equalizing maps, Michael T. Gastner and M. E. J. Newman, Proc. Natl. Acad. Sci. USA 101, 7499-7504 (2004)
}
\seealso{
\link{quick.carto} \link{mesh}
}
\keyword{cartogram}

