% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mesh}
\alias{mesh}
\title{Visual representation of cartogram transformation}
\usage{
mesh(sldf, add = FALSE, col = "red")
}
\arguments{
\item{add}{Whether to add the graphic to an existing plot}

\item{col}{Colour of the graphic}

\item{spdf}{A \code{SpatialPolygonsDataFrame}  created via the \code{quick.carto} function or similar}
}
\description{
Given a transformed Spatial object draws the warping grid
}
\examples{
# Load an example SpatialPolygonsDataFrame from GISTools
data(georgia)
# Create the cartogram.  TotPop90 contains 1990 county populations
# "georgia2" is used as it has a plane projection coordinate system
georgia.carto <- quick.carto(georgia2,georgia2$TotPop90)
# Draw the mesh
mesh(georgia.carto)
# Add the transformed map
plot(georgia.carto,add=TRUE)
}
\references{
A Diffusion-based method for producing density equalizing maps, Michael T. Gastner and M. E. J. Newman, Proc. Natl. Acad. Sci. USA 101, 7499-7504 (2004)
}
\seealso{
\link{quick.carto} \link{dispersion}
}
\keyword{cartogram}

