% Generated by roxygen2 (4.0.2): do not edit by hand
\name{quick.carto}
\alias{quick.carto}
\title{Create a cartogram SpatialPolygonsDataFrame}
\usage{
quick.carto(spdf, v, extend = 0.05, res = 128, index = 1, thresh = 0)
}
\arguments{
\item{spdf}{A \code{SpatialPolygonsDataFrame}}

\item{v}{The variable providing the values - typically population counts}

\item{extend}{The extent beyond the limits of the \code{SpatialPolygonsDataFrame} to estimate the warping as a proportion of span}

\item{res}{The resolution of the grid for the warping}

\item{index}{Index of data item to be used for cartogram - only used if \code{v} is missing}

\item{thresh}{Lowest density value allowed - will replace pixels with \code{max(value,thresh)}}
}
\value{
A \code{SpatialPolygonsDataFrame} containing the cartogram.  Also has an
attribute - \code{attr(spdf,'mesh')} - containing information about the cartogram
warping grid
}
\description{
Given a \code{SpatialPolygonsDataFrame} and a value associated with each polygon
creates a new \code{SpatialPolygonsDataFrame} which is a cartogram of the input
whose expansion factors are based on the value divided by the polygon area.
}
\examples{
data(georgia)
georgia.carto <- quick.carto(georgia,georgia2$TotPop90)
par(mfrow=c(1,2))
plot(georgia2)
plot(georgia.carto)
}
\references{
A Diffusion-based method for producing density equalizing maps, Michael T. Gastner and M. E. J. Newman, Proc. Natl. Acad. Sci. USA 101, 7499-7504 (2004)
}
\seealso{
\link{mesh}
}
\keyword{cartogram}

