% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quick.carto.transform.R
\name{quick.carto.transform}
\alias{quick.carto.transform}
\title{Create a warping function to apply to Spatial* and Spatial*DataFrame objects}
\usage{
quick.carto.transform(poly)
}
\arguments{
\item{poly}{a \code{SpatialPolygonsDataFrame} created by \code{quick.carto}}
}
\value{
a function taking a \code{Spatial*} or \code{Spatial*DataFrame} object,  returning a warped version of the object
}
\description{
Given an object created by \code{quick.carto},  create a cartogram transform.
The returned value is a function that performs the transform.
}
\examples{
# Make a cartogram for Newhaven census blocks
require(GISTools)
data(newhaven)
blocks.carto <- quick.carto(blocks,blocks$POP1990)
# Create the transform function
to.carto <- quick.carto.transform(blocks.carto)
# Plot the blocks carto
plot(blocks.carto)
# Add roads,  transformed to cartogram space
plot(to.carto(roads),add=TRUE,col='lightgrey')
# Add forced entry residential burglaries, transformed to cartogram space
plot(to.carto(burgres.f),add=TRUE,col='red',pch=16)
}
\references{
A Diffusion-based method for producing density equalizing maps, Michael T. Gastner and M. E. J. Newman, Proc. Natl. Acad. Sci. USA 101, 7499-7504 (2004)
}
\seealso{
Other transformers: \code{\link{carto.transform}};
  \code{\link{warp.lines}}; \code{\link{warp.points}};
  \code{\link{warp.polys}}
}
\keyword{cartogram}

