% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/warp.lines.R
\name{warp.lines}
\alias{warp.lines}
\title{Warp a SpatialLines or SpatialLinesDataFrame object according a to cartogram transform}
\usage{
warp.lines(sldf, warper, prec = 0.5)
}
\arguments{
\item{sldf}{a \code{SpatialLines} or \code{SpatialLinesDataFrame} object to be transformed}

\item{warper}{a \code{SpatialPolygonsDataFrame} or \code{SpatialPolygons} object resulting from a previous \code{quick.carto} evocation}

\item{prec}{The precision when carrying out interpolation, in units of the warping grid - 0.5 works well}
}
\value{
a \code{SpatialLines} or \code{SpatialLinesDataFrame} object resulting from the transformation.
}
\description{
Given a \code{SpatialLines} or \code{SpatialLinesDataFrame} object,  apply a cartogram transform
according to the transform grid of a \code{SpatialPolygonsDataFrame} created via \code{quick.carto}
The returned values is the transformed \code{SpatialLines} or \code{SpatialLinesDataFrame}.
}
\examples{
# Make a cartogram for Georgia counties
data(georgia)
georgia.carto <- quick.carto(georgia2,georgia2$TotPop90)

# Create a line from southwest to northeast corners,  with 100 segments
xl <- seq(939220.7,1419424,l=100)
yl <- seq(905508,1405900,l=100)
aline <- readWKT(paste("LINESTRING(",paste(apply(cbind(xl,yl),1,function(x) paste(x[1],x[2])),collapse=","),")"))

#  Transform it
aline.carto <- warp.lines(aline,georgia.carto)

# Plot the original line
par(mfrow=c(2,1),mar=c(0.5,0.5,3,0.5))
plot(georgia2,border='grey'); plot(aline, add=T, col='red')
title('Original projection')

# Plot in cartogram space
plot(georgia.carto,border='grey'); plot(aline.carto, add=T, col='red')
title('Cartogram projection')
}
\references{
A Diffusion-based method for producing density equalizing maps, Michael T. Gastner and M. E. J. Newman, Proc. Natl. Acad. Sci. USA 101, 7499-7504 (2004)
}
\seealso{
Other transformers: \code{\link{carto.transform}};
  \code{\link{quick.carto.transform}};
  \code{\link{warp.points}}; \code{\link{warp.polys}}
}
\keyword{cartogram}

