% Generated by roxygen2 (4.0.2): do not edit by hand
\name{warp.polys}
\alias{warp.polys}
\title{Warp a SpatialPolygons or SpatialPolygonsDataFrame object according a to cartogram transform}
\usage{
warp.polys(spdf, warper, prec = 0.5)
}
\arguments{
\item{spdf}{a \code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} object to be transformed}

\item{warper}{a \code{SpatialPolygonsDataFrame} or \code{SpatialPolygons} object resulting from a previous \code{quick.carto} evocation}

\item{prec}{The precision when carrying out interpolation, in units of the warping grid - 0.5 works well}
}
\value{
a \code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} object resulting from the transformation.
}
\description{
Given a \code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} object,  apply a cartogram transform
according to the transform grid of a \code{SpatialPolygonsDataFrame} created via \code{quick.carto}
The returned values is the transformed \code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}.
}
\examples{
# Make a cartogram for Georgia counties
data(georgia)
georgia.carto <- quick.carto(georgia2,georgia2$TotPop90)

# Create a pseudocircle with 100 segments
xl <- cos(seq(0,2*pi,l=100))*100000 + 1239348
yl <- sin(seq(0,2*pi,l=100))*100000 + 1093155
circ <- readWKT(paste("MULTIPOLYGON(((",paste(apply(cbind(xl,yl),1,function(x) sprintf("\%7.0f \%7.0f",x[1],x[2])),collapse=","),")))"))
#' #  Transform it
circ.carto <- warp.polys(circ,georgia.carto)

# Plot the original circle
par(mfrow=c(2,1),mar=c(0.5,0.5,3,0.5))
plot(georgia2,border='grey'); plot(circ, add=T, col=rgb(1,0,0,0.4), border=NA)
title('Original projection')

# Plot in cartogram space
plot(georgia.carto,border='grey'); plot(circ.carto, add=T, col=rgb(1,0,0,0.4), border=NA)
title('Cartogram projection')
}
\references{
A Diffusion-based method for producing density equalizing maps, Michael T. Gastner and M. E. J. Newman, Proc. Natl. Acad. Sci. USA 101, 7499-7504 (2004)
}
\seealso{
Other transformers: \code{\link{carto.transform}};
  \code{\link{quick.carto.transform}};
  \code{\link{warp.lines}}; \code{\link{warp.points}}
}
\keyword{cartogram}

